refactor(payment, reminder, schedule, treatment, user): standradize API

+++ src/main/java/com/emmkay/infertility_system_api/modules/treatment/dto/request/TreatmentServiceCreateRequest.java
@@ -27,6 +27,6 @@ public class TreatmentServiceCreateRequest {
     int duration;
 
     @NotNull(message = "{validation.required}")
-    int treatmentTypeId;
+    Long treatmentTypeId;
 
 }
